refactor(k8s-version): Move darling code into module
